Removed default Lat Long from formik values and merged utility pages.
