+++ src/components/SignUpForm.jsx
@@ -10,8 +10,7 @@ import { IsFormInComplete, SetFieldValues } from "../utilities/FormHelpers";
 import { IsPhoneNumberValid } from '@baltimorecounty/validation';
 import SeButton from "./SeButton";
 import { GoHome, Go, Routes } from "../Routing";
-import { getAlertMessage, resetAlerts } from "../utilities/AlertHelper";
-import { AlertAtPage } from "../utilities/AlertHelpers";
+import { AlertAtPage, getAlertMessage, resetAlerts } from "../utilities/AlertHelpers";
 const CreateAccount = (props, routeProps) => {
 
 	const { Tabs, SignUpPage, shouldDisableForm } = props.values;
