+++ src/components/withFormik.jsx
@@ -47,8 +47,8 @@ export default withFormik({
 		location: '',
 		describeTheProblem: '',
 		MapDefaults: [''],
-		Latitude: 39.40037792,
-		Longitude: -76.60651907,
+		Latitude: '',
+		Longitude: '',
 		ShowErrorMsg: '',
 		contactID: '',
 		streetAddress: '',
